Use relative paths in the tarball, also a shorter path
No need to encode the entire path to the user's home directory in to the
paths of the files in the tarball. All that matters is the tempdir. This
ensures that the pathing is relative, short and sane. Also adds an exit
if tempdir already exists (which given the timestamp seems weird that it
would) because you wouldn't want to risk mixing old data with new data
in a single tarball.