Remove /etc/services grep from the script
Every year, we use this script as part of our assessment, and every year
I get questions about all the services we're running.  "Why are we
running FTP and Telnet?" "We're not." It's the output of this script
that throws the assessor.

Specifically, the `grep 21 /etc/services` line.  This produces a ton of
irrelevant information. `/etc/services` is a port/protocol
"registration" for the local system.  It helps other tools like
`netstat`, `lsof`, and `ss` provide human readable names to their output
of ports.  It does *not* have anything to do with running or configured
services.

`grep 21 /etc/services` also is an incorrect use of `grep` as it matches
all lines containing '21', whether or not it has anything to do with
port 21.

I'm removing these two `grep .. /etc/services` lines from the script as
they are confusing and unhelpful and do not provide any PCI relevant
information.